+++ aio/scripts/deploy-to-firebase/index.mjs
@@ -5,6 +5,7 @@
 
 import sh from 'shelljs';
 import {fileURLToPath} from 'url';
+import post from './post-deploy-actions.mjs';
 import pre from './pre-deploy-actions.mjs';
 import u from './utils.mjs';
 
@@ -100,7 +101,7 @@ function computeDeploymentsInfo(
       siteId: 'next-angular-io-site',
       deployedUrl: 'https://next.angular.io/',
       preDeployActions: [pre.build, pre.checkPayloadSize],
-      postDeployActions: [testPwaScore],
+      postDeployActions: [post.testPwaScore],
     },
     rc: {
       name: 'rc',
@@ -110,7 +111,7 @@ function computeDeploymentsInfo(
       siteId: 'rc-angular-io-site',
       deployedUrl: 'https://rc.angular.io/',
       preDeployActions: [pre.build, pre.checkPayloadSize],
-      postDeployActions: [testPwaScore],
+      postDeployActions: [post.testPwaScore],
     },
     stable: {
       name: 'stable',
@@ -120,7 +121,7 @@ function computeDeploymentsInfo(
       siteId: `v${currentBranchMajorVersion}-angular-io-site`,
       deployedUrl: 'https://angular.io/',
       preDeployActions: [pre.build, pre.checkPayloadSize],
-      postDeployActions: [testPwaScore],
+      postDeployActions: [post.testPwaScore],
     },
     archive: {
       name: 'archive',
@@ -130,7 +131,7 @@ function computeDeploymentsInfo(
       siteId: `v${currentBranchMajorVersion}-angular-io-site`,
       deployedUrl: `https://v${currentBranchMajorVersion}.angular.io/`,
       preDeployActions: [pre.build, pre.checkPayloadSize],
-      postDeployActions: [testPwaScore],
+      postDeployActions: [post.testPwaScore],
     },
 
     // SECONDARY DEPLOY TARGETS
@@ -151,7 +152,7 @@ function computeDeploymentsInfo(
       siteId: 'rc-angular-io-site',
       deployedUrl: 'https://rc.angular.io/',
       preDeployActions: [pre.disableServiceWorker, pre.redirectToAngularIo],
-      postDeployActions: [testNoActiveRcDeployment],
+      postDeployActions: [post.testNoActiveRcDeployment],
     },
   };
 
@@ -277,52 +278,6 @@ function serializeActions(actions) {
 
 function skipDeployment(reason) {
   return {name: 'skipped', type: 'skipped', reason};
-}
-
-function testNoActiveRcDeployment({deployedUrl}) {
-  u.logSectionHeader(
-      'Verify deployed RC version redirects to stable (and disables old ServiceWorker).');
-
-  const deployedOrigin = deployedUrl.replace(/\/$/, '');
-
-  // Ensure a request for `ngsw.json` returns 404.
-  const ngswJsonUrl = `${deployedOrigin}/ngsw.json`;
-  const ngswJsonScript = `https.get('${ngswJsonUrl}', res => console.log(res.statusCode))`;
-  const ngswJsonActualStatusCode =
-      sh.exec(`node --eval "${ngswJsonScript}"`, {silent: true}).trim();
-  const ngswJsonExpectedStatusCode = '404';
-
-  if (ngswJsonActualStatusCode !== ngswJsonExpectedStatusCode) {
-    throw new Error(
-        `Expected '${ngswJsonUrl}' to return a status code of '${ngswJsonExpectedStatusCode}', ` +
-        `but it returned '${ngswJsonActualStatusCode}'.`);
-  }
-
-  // Ensure a request for `foo/bar` is redirected to `https://angular.io/foo/bar`.
-  const fooBarUrl = `${deployedOrigin}/foo/bar?baz=qux`;
-  const fooBarScript =
-      `https.get('${fooBarUrl}', res => console.log(res.statusCode, res.headers.location))`;
-  const [fooBarActualStatusCode, fooBarActualRedirectUrl] =
-      sh.exec(`node --eval "${fooBarScript}"`, {silent: true}).trim().split(' ');
-  const fooBarExpectedStatusCode = '302';
-  const fooBarExpectedRedirectUrl = 'https://angular.io/foo/bar?baz=qux';
-
-  if (fooBarActualStatusCode !== fooBarExpectedStatusCode) {
-    throw new Error(
-        `Expected '${fooBarUrl}' to return a status code of '${fooBarExpectedStatusCode}', but ` +
-        `it returned '${fooBarActualStatusCode}'.`);
-  } else if (fooBarActualRedirectUrl !== fooBarExpectedRedirectUrl) {
-    const actualBehavior = (fooBarActualRedirectUrl === 'undefined') ?
-      'not redirected' : `redirected to '${fooBarActualRedirectUrl}'`;
-    throw new Error(
-        `Expected '${fooBarUrl}' to be redirected to '${fooBarExpectedRedirectUrl}', but it was ` +
-        `${actualBehavior}.`);
-  }
-}
-
-function testPwaScore({deployedUrl, minPwaScore}) {
-  u.logSectionHeader('Run PWA-score tests.');
-  u.yarn(`test-pwa-score "${deployedUrl}" "${minPwaScore}"`);
 }
 
 function validateDeploymentsInfo(deploymentsList) {
